Take ysh off nightly and set it to use stable
The `crate` modifier was the only nightly feature we used that's
a syntactic nicety but can be done with `pub(crate)`. This takes ysh
off nightly by using a rust-toolchain file for stable, supresses the
deprecation warnings added since the last update, and changes the
modifier. Using stable is in the long term a better idea not only for
adoption, but means we can release at anytime without having to worry
about people who can't use nightly.